+++ src/main/java/com/ClinicaDelCalzado_BackEnd/repository/userAdmin/IAdministratorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IAdministratorRepository  extends JpaRepository<Administrator, Integer> {
+public interface IAdministratorRepository  extends JpaRepository<Administrator, Long> {
 }
